renaming elements and adding documentation
At this stage the MVVM design pattern is in place. The view is not
handling all the logic but rather delegates to the view-model the
task of prompting the user to choose an image file. A service has
been added to support the view-model, and possibly other future
view-models, with this task. The service is not aware of what the
view wants, which is actually the responsibility of the view-model.
The service merely helps in providing the user with a tool to
navigate the file system for selecting an image. Once a selection is
made, the service returns to the caller, i.e. the view-model, a
model object that represents the image. Then, the view-model updates
its properties which results in notifications about their values
having been changed. The view which has elements binding to these
properties acts upon these notifications to obtain and display the
new values.

A MVVM library (Microsoft.Toolkit.Mvvm) has been installed to
minimize reinventing the wheel on certain aspects.