changed view-model to receive service instead of instantiating it
More towards achieving a loosely coupled architecture is introducing
dependency injection technique to the view-model. The view-model is
no more holding ownership of the service instance. The service is
created by the view which is passed as a parameter to the view-model
constructor. Also, the view-model can now be unit tested easily when
needed by injecting its constructor with a mock-up service.