dependency injection to view-model is delegated to the IoC container
To construct the view-model, its dependencies need to be passed to
its constructor. This used to be done by the view. To remove this
burden, the task is now delegated to a IoC container. The view
simply asks this container to provide the view-model instance whose
dependencies are resolved by the container. For this, the IoC
container is initialized at application launch, by registering
the services it needs to provide on demand.