moving the content of the main view-model to separate view-models
The main page includes two pages both using the main view-model. Now
each page has its own view-model with only the functionality and
data needed by the page. Because one view-model needs to provide the
model for the second view-model without the two knowing about each
other, a singleton class was added that has a property for holding
this model. The view-model in charge of getting the image sets this
property. The view-model in charge of providing image details to the
view is listening to value changes of the property to make necessary
updates for the view.