using messenger to exchange messages between view-models
The installed MVVM toolkit provides a messenger implementation that
helps in communicating messages between different modules and keep
them loosely coupled. This is more useful and convenient than using
the AppState to hold the image model that would be accessible by all
the view-models. The view-model that obtains the image selected by
user sends a message whenever the image is available. The view-model
that is meant for only displaying the image details receives the
message containing the new image model and does necessary view
updates.